requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
paddlepaddle==2.6.2
paddleocr==2.7.3
Pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.1
